feat: Update vote routes to match new contract and create vote controller
